feat(cme-form): Add Text Splitter
1. The version in `frontend/package.json` is updated from "0.0.0" to
"0.1.1", as there have been some changes or updates to the frontend
package.

2. The version in `package.json` is updated from "0.25.0" to "0.26.1",
as there have been some changes or updates to the extension package.

3. In `frontend/src/components/cme-form-view/FormBuilder.ts`

   1. a new import statement is added:
`import { splitTextEventHandler } from '../../utils/splitText';`. This
means that a new function `splitTextEventHandler` is imported from the
`splitText` module in the `utils` directory.
   2. Inside the `class FormBuilder`, a new event listener is added to
the `vscode-inputbox` element with the `@vsc-input` event triggering the
`splitTextEventHandler` function. This event handler is tied to the
splitting of text entered in the `vscode-inputbox`.

4. In `frontend/src/components/cme-form-view/TemplateCompiler.ts`

   1. a new import statement is added:
`import splitText from '../../utils/splitText';`. This means that the
`splitText` function from the `utils` directory is imported.
   2. Inside the `class TemplateCompiler`, the `splitText` function is
called with the token value and then assigned to the `value` variable.

5. In `frontend/src/utils/splitText.ts`

   1. a new function `splitTextInInputBox` is defined. This function
splits a string into chunks of 72 characters while preserving line
breaks and punctuation, and replaces the `value` of a `VscodeInputbox`
element with the split result.
   2. Another function `splitText` is defined, which splits a string
into chunks of 72 characters while preserving line breaks and
punctuation. It uses regular expressions to split the string based on
the character limit and handle punctuations.
   3.  In the same file, a new function `splitTextEventHandler` is
defined. This function is an event handler for `VscodeInputbox`
elements. If the target element has a data attribute called "name" with
a value of "body", it calls the `splitTextInInputBox` function with the
event details and target element as arguments.

6. In `frontend/tsconfig.json`, a new entry is added to the `"lib"`
array: `"ES2021.String"`. This indicates that the code is now targeting
the ES2021 String API.