feat: Add ability to parse commit messages
This commit adds a new function `parseCommitMessage` in the
`Tokenizer.ts` file which parses a commit message string and returns an
object with the parsed values. This function is used in the
`cme-form-view.ts` file to extract token values from the commit message
and fill a form with the token values.

Changes include:
- Added `parseCommitMessage` function in `Tokenizer.ts`
- Updated `cme-form-view.ts` to use `parseCommitMessage` to extract
token values from commit message
- Updated `EditorController.ts` to handle the
'copyToExtensionMessageBox' command and send the SCM input box value to
the extension
- Added event listener and method `_handleCopyFromSCMInputBox` in
`cme-form-view.ts` to copy token values from SCM input box
- Updated `package.json` with new version
- Added `showCOMMIT_EDITMSG` method in `GitService.ts` to show the
commit message editor