fix: git API only tags HEAD, not specific commit
The Commit interface in the git.d.ts file has been modified to include a
new "refNames" property. This property contains an array of strings
representing the reference names associated with the commit.

Additionally, changes have been made to the VersionGitTagCommand. A
check has been added to skip tagging if the commit's refNames include
the tag. This ensures that tags are only added to commits without
existing reference names.

Furthermore, the GitService class has been refactored to use the Visual
Studio Code built-in terminal for creating a tag at a specific commit.
This is because the native git API only tags HEAD, which is not always
the commit that the user wants to tag.