feat: Add method to get number of commits
This commit adds a new method `getNumberOfCommits` to the `GitService`
class. This method calculates the number of commits in a given
repository by executing the command `git rev-list --count HEAD`. The
result is returned as a number.

This method is used in the `VersionGitTagCommand` to determine the
number of recent commits for retrieving commit messages.