Change TagVersionRegexes to TagVersion and add to configuration settings obejct
This change updates the code to replace the `TagVersionRegexes` type
with `TagVersion` type in the global declaration file. It also modifies
the `ExtensionConfig` interface to use `TagVersion` instead of
`TagVersionRegexes` in the `gitTagVersion` property. Additionally, the
`TagVersion` type is updated to include properties for `prefix`, `body`,
`description`, and `files`. The `createInitialState` function, the
`cme-form-view.test.ts` test file, `cme-editor-page.test.ts` test file,
and the `package.json` file all reflect these code changes. In the
`VersionGitTagCommand` file, the `tagTemplate` is changed to use the
updated `gitTagVersion` configuration. The `VersionGitTagCommand` file
also imports the `TagVersion` type from the global declaration file.