feat: Add GitTagVersionCompletionProvider code changes
This commit adds code changes related to the
GitTagVersionCompletionProvider. It includes the addition of a new type
and the modification of the constructor. The provideCompletionItems
method has been updated to include additional logic. The addTagSnippet
method now accepts an optional version parameter. Two new methods,
getHighestExistingVersion and compareVersions, have been added. The
GitService class has also been updated with the addition of the
unstageAllChanges method.