feat: improve terminal switching behavior
Improve the behavior of the `SwitchToTerminalByNameCommand` to better
handle terminal switching in different workspace folders.

* The command now takes into account the current workspace folder when
switching to a terminal
* If the current workspace folder is the same as the terminal's working
directory, the command will reuse the existing terminal
* If the current workspace folder is different from the terminal's
working directory, the command will create a new terminal with the
correct working directory

This change is necessary to improve the user experience when working
with multiple workspace folders and terminals. By reusing existing
terminals when possible and creating new terminals with the correct
working directory, the command reduces the number of unnecessary
terminals and makes it easier for users to manage their terminals.

This commit is part of a larger effort to improve the terminal
management features of the extension, making it easier for users to work
with terminals in different workspace folders.